visual-analytics-engine/requirements.txt
aiocache==0.11.1
aiohttp==3.7.4.post0
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
decorator==4.4.2
fastapi==0.67.0
greenlet==1.1.0
gssapi==1.6.14
h11==0.12.0
hiredis==2.0.0
idna==3.2
importlib-metadata==4.6.1
JayDeBeApi==1.2.3
joblib==1.0.1
JPype1==1.3.0
llvmlite==0.36.0
msgpack==1.0.2
multidict==5.1.0
networkx==2.5.1
nptyping==1.4.2
numba==0.53.1
numpy==1.21.1
pandas==1.3.1
phoenixdb==1.0.1
protobuf==3.17.3
psycopg2==2.9.1
pydantic==1.8.2
pynndescent==0.5.4
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests-gssapi==1.2.3
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.22
starlette==0.14.2
strsim==0.0.3
threadpoolctl==2.2.0
typing-extensions==3.10.0.0
typish==1.9.2
ujson==4.0.2
umap-learn==0.5.1
urllib3==1.26.6
uvicorn==0.14.0
yarl==1.6.3
zipp==3.5.0
